feat(core & adapter): :sparkles: Add new features and fix bugs
A new decorator called '@LibClient' was added, and now a class is responsible for obtaining the return of a command, to convert it into a response for the user. The possibility of dividing the slash commands into global ones was also added, so that when they are going to be published, only publish those that have the global set to true, apart from also being able to choose which commands to publish specifically.